move 'on' back to routing
