+++ Sources/Routing/AsyncRouter.swift
@@ -1,6 +1,23 @@
 import Async
 import HTTP
 
+/// Capable of registering async routes.
+extension Router {
+    /// Registers a route handler at the supplied path.
+    @discardableResult
+    public func on<F: FutureType>(
+        _ method: Method,
+        to path: [PathComponent],
+        use closure: @escaping BasicResponder<F>.Closure
+        ) -> Route where F.Expectation: ResponseRepresentable {
+        let responder = BasicResponder(closure: closure)
+        let route = Route(method: method, path: path, responder: responder)
+        self.register(route: route)
+        
+        return route
+    }
+}
+
 /// A basic, closure-based responder.
 public struct BasicResponder<F: FutureType>: Responder where F.Expectation: ResponseRepresentable {
     /// Responder closure
@@ -21,4 +38,3 @@ public struct BasicResponder<F: FutureType>: Responder where F.Expectation: Resp
         }
     }
 }
-
