+++ Sources/Vapor/Routing.swift
@@ -21,23 +21,6 @@ public struct RouterResponder: Responder {
     }
 }
 
-/// Capable of registering async routes.
-extension Router {
-    /// Registers a route handler at the supplied path.
-    @discardableResult
-    public func on<F: FutureType>(
-        _ method: Method,
-        to path: [PathComponent],
-        use closure: @escaping BasicResponder<F>.Closure
-        ) -> Route where F.Expectation: ResponseRepresentable {
-        let responder = BasicResponder(closure: closure)
-        let route = Route(method: method, path: path, responder: responder)
-        self.register(route: route)
-        
-        return route
-    }
-}
-
 extension Router {
     /// Creates an Async `Route` at the provided path using the `GET` method.
     @discardableResult
